Use fixed int width

+++ tests/domains_test.py
@@ -30,7 +30,7 @@ def test_find_domains_length_1():
 def test_find_domains_length_2():
     data = sc.ones(dims=['x'], shape=[2], dtype='int64')
     da = sc.DataArray(data=data, coords={'x': sc.array(dims=['x'], values=[0, 1, 2])})
-    expected = sc.DataArray(data=sc.array(dims=['x'], values=[1]),
+    expected = sc.DataArray(data=sc.array(dims=['x'], values=[1], dtype='int64'),
                             coords={'x': sc.array(dims=['x'], values=[0, 2])})
     assert sc.identical(find_domains(da), expected)
 
@@ -45,7 +45,7 @@ def test_find_domains_constant():
     data = sc.ones(dims=['x'], shape=[9], dtype='int64')
     da = sc.DataArray(data=data, coords={'x': x})
 
-    expected = sc.DataArray(data=sc.array(dims=['x'], values=[1]),
+    expected = sc.DataArray(data=sc.array(dims=['x'], values=[1], dtype='int64'),
                             coords={'x': sc.array(dims=['x'], values=[0, 9])})
     assert sc.identical(find_domains(da), expected)
